pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "Kurigram"
dynamic = ["version"]
description = "Elegant, modern and asynchronous Telegram MTProto API framework in Python for users and bots"
readme = "README.md"
requires-python = ">=3.8"
authors = [{ name = "Dan", email = "dan@pyrogram.org" }]
maintainers = [{ name = "KurimuzonAkuma" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Internet",
    "Topic :: Communications",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]
dependencies = ["pyaes<=1.6.1", "pysocks<=1.7.1"]
license = "LGPL-3.0-or-later"
license-files = ["COPYING", "COPYING.lesser"]
keywords = [
    "telegram",
    "chat",
    "messenger",
    "mtproto",
    "api",
    "client",
    "library",
    "python",
]

[project.urls]
Homepage = "https://kurigram.icu"
Documentation = "https://docs.kurigram.icu"
Source = "https://github.com/KurimuzonAkuma/pyrogram"
Issues = "https://github.com/KurimuzonAkuma/pyrogram/issues"
Community = "https://t.me/kurigram_chat"

[project.optional-dependencies]
dev = [
    "keyring<=25.7.0",
    "hatch<=1.16.2",
    "pytest<=9.0.2",
    "pytest-asyncio<=1.3.0",
    "pytest-cov<=7.0.0",
    "twine<=6.2.0",
]

docs = [
    "Sphinx<=9.0.4",
    "shibuya<=2025.12.19",
    "sphinx-autobuild<=2025.8.25",
    "sphinx-copybutton<=0.5.2",
    "sphinx-iconify<=0.2.1",
    "sphinx-design<=0.6.1",
    "pygments<=2.19.2",
]

fast = [
    "tgcrypto<=1.2.5",
    "uvloop<=0.21.0; sys_platform == 'darwin' or sys_platform == 'linux'",
]

[tool.hatch.version]
path = "pyrogram/__init__.py"

[tool.hatch.build.hooks.custom]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.sdist]
exclude = [
    ".github/",
    "docs/",
    "tests/",
    ".gitignore",
]

[tool.hatch.build.targets.wheel]
ignore-vcs = true
packages = ["pyrogram"]
